Reorganizing single task background concurrency tools
