+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/features/centralconfig/initializer/CentralConfigurationInitializer.java
@@ -22,9 +22,9 @@ import androidx.annotation.VisibleForTesting;
 
 import co.elastic.apm.android.sdk.internal.features.centralconfig.CentralConfigurationManager;
 import co.elastic.apm.android.sdk.internal.features.centralconfig.poll.ConfigurationPollManager;
-import co.elastic.apm.android.sdk.internal.utilities.concurrency.BackgroundExecutor;
-import co.elastic.apm.android.sdk.internal.utilities.concurrency.Result;
-import co.elastic.apm.android.sdk.internal.utilities.concurrency.impl.SimpleBackgroundExecutor;
+import co.elastic.apm.android.sdk.internal.utilities.concurrency.singletask.BackgroundExecutor;
+import co.elastic.apm.android.sdk.internal.utilities.concurrency.singletask.Result;
+import co.elastic.apm.android.sdk.internal.utilities.concurrency.singletask.impl.SimpleBackgroundExecutor;
 
 public final class CentralConfigurationInitializer implements BackgroundExecutor.Callback<Integer> {
     private final BackgroundExecutor executor;
