+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/utilities/concurrency/singletask/BackgroundExecutor.java
@@ -16,23 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.utilities.concurrency.impl;
+package co.elastic.apm.android.sdk.internal.utilities.concurrency.singletask;
 
-import co.elastic.apm.android.sdk.internal.utilities.concurrency.BackgroundExecutor;
-import co.elastic.apm.android.sdk.internal.utilities.concurrency.BackgroundWork;
-import co.elastic.apm.android.sdk.internal.utilities.concurrency.Result;
+public interface BackgroundExecutor {
 
-public class SimpleBackgroundExecutor implements BackgroundExecutor {
+    <T> void execute(BackgroundWork<T> work, Callback<T> callback);
 
-    @Override
-    public <T> void execute(BackgroundWork<T> work, Callback<T> callback) {
-        new Thread(() -> {
-            try {
-                T result = work.execute();
-                callback.onFinish(Result.success(result));
-            } catch (Throwable t) {
-                callback.onFinish(Result.error(t));
-            }
-        }).start();
+    interface Callback<T> {
+        void onFinish(Result<T> result);
     }
 }
